requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
python-decouple>=3.8
django-filter>=23.5
psycopg2-binary>=2.9
drf-spectacular>=0.27
gunicorn>=22.0
whitenoise>=6.7
torch>=2.0
torchvision>=0.15
celery>=5.3
redis>=5.0
django-ratelimit>=4.1
torchaudio>=2.0
transformers>=4.30
librosa
scipy
